# openapi.yml
openapi: 3.0.3
info:
  title: Community Activity API
  description: API for managing users, activities and postcode coordinates.
  version: 1.0.0

servers:
  - url: https://api.example.com/v1
    description: Production

tags:
  - name: Users
  - name: Activities
  - name: Tasks
  - name: Metrics
  - name: Coordinates

paths:
  /users:
    get:
      tags: [Users]
      summary: Get list of users
      operationId: listUsers
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'

  /users/new:
    post:
      tags: [Users]
      summary: Create a user
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserInput'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'

  /users/{user_id}:
    parameters:
      - $ref: '#/components/parameters/UserId'
    get:
      tags: [Users]
      summary: Get a single user
      operationId: getUser
      responses:
        '200':
          description: The user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Users]
      summary: Update a user
      operationId: updateUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserInput'
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Users]
      summary: Delete a user
      operationId: deleteUser
      responses:
        '204':
          description: User deleted
        '404':
          $ref: '#/components/responses/NotFound'

  /activities:
    get:
      tags: [Activities]
      summary: Get activities
      description: >
        Returns activities, optionally filtered by user, task and/or date.
        Filters are combined with AND; with no filters, all activities are returned.
      operationId: listActivities
      parameters:
        - name: user_id
          in: query
          required: false
          description: Only return activities for this user
          schema:
            type: integer
            format: int64
            example: 1
        - name: task_id
          in: query
          required: false
          description: Only return activities for this task
          schema:
            type: integer
            format: int64
            example: 42
        - name: date
          in: query
          required: false
          description: Only return activities on this date (YYYY-MM-DD)
          schema:
            type: string
            format: date
            example: '2026-10-03'
      responses:
        '200':
          description: List of activities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Activity'
        '400':
          $ref: '#/components/responses/BadRequest'

  /activities/{date}/{user_id}/{task_id}:
    parameters:
      - name: date
        in: path
        required: true
        description: Date of the activity (YYYY-MM-DD)
        schema:
          type: string
          format: date
          example: '2026-10-03'
      - $ref: '#/components/parameters/UserId'
      - name: task_id
        in: path
        required: true
        description: ID of the task
        schema:
          type: integer
          format: int64
          example: 42
    post:
      tags: [Activities]
      summary: Record an activity
      description: >
        Records that the user completed the task on the date. Points are taken
        from the task. (user_id, task_id, date) is a joint key, so a duplicate
        returns 409.
      operationId: createActivity
      responses:
        '201':
          description: Activity created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Activity'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
    delete:
      tags: [Activities]
      summary: Delete an activity
      operationId: deleteActivity
      responses:
        '204':
          description: Activity deleted
        '404':
          $ref: '#/components/responses/NotFound'

  /tasks:
    get:
      tags: [Tasks]
      summary: Get list of tasks
      operationId: listTasks
      responses:
        '200':
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'

  /metrics:
    get:
      tags: [Metrics]
      summary: Get total points for each postcode
      description: >
        Returns the total points aggregated over all activities recorded
        by users in each postcode.
      operationId: listPostcodeMetrics
      responses:
        '200':
          description: Total points per postcode
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PostcodeMetric'

  /metrics/{postcode}:
    get:
      tags: [Metrics]
      summary: Get metrics for a postcode
      description: >
        Returns the postcode's carbon intensity and air quality, plus the
        users in the postcode with their points (ordered highest first).
      operationId: getPostcodeUserMetrics
      parameters:
        - name: postcode
          in: path
          required: true
          description: Postcode to get user points for (e.g. EH1 1YZ)
          schema:
            type: string
            example: EH1 1YZ
      responses:
        '200':
          description: Metrics for the postcode
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostcodeDetailMetric'
        '404':
          $ref: '#/components/responses/NotFound'

  /coordinates:
    get:
      tags: [Coordinates]
      summary: Get latitude and longitude for all postcodes
      operationId: listCoordinates
      responses:
        '200':
          description: List of postcode coordinates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Coordinates'

  /coordinates/{postcode}:
    get:
      tags: [Coordinates]
      summary: Get latitude and longitude for a single postcode
      operationId: getCoordinates
      parameters:
        - name: postcode
          in: path
          required: true
          description: Postcode to look up (e.g. EH1 1YZ)
          schema:
            type: string
            example: EH1 1YZ
      responses:
        '200':
          description: Coordinates for the postcode
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Coordinates'
        '404':
          $ref: '#/components/responses/NotFound'

components:
  parameters:
    UserId:
      name: user_id
      in: path
      required: true
      description: ID of the user
      schema:
        type: integer
        format: int64
        example: 1

  schemas:
    User:
      type: object
      required: [id, name, email, postcode]
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: Jane Smith
        email:
          type: string
          format: email
          example: jane@example.com
        postcode:
          type: string
          example: EH1 1YZ

    UserInput:
      type: object
      required: [name, email, postcode]
      properties:
        name:
          type: string
          example: Jane Smith
        email:
          type: string
          format: email
          example: jane@example.com
        postcode:
          type: string
          example: EH1 1YZ

    Activity:
      type: object
      description: Keyed by (user_id, task_id, date).
      required: [task_id, user_id, points, date]
      properties:
        task_id:
          type: integer
          format: int64
          example: 42
        user_id:
          type: integer
          format: int64
          example: 1
        points:
          type: integer
          example: 10
        date:
          type: string
          format: date
          example: '2026-10-03'

    Task:
      type: object
      required: [id, name, description, points]
      properties:
        id:
          type: integer
          format: int64
          example: 42
        name:
          type: string
          example: Park litter pick
        description:
          type: string
          example: Litter pick in the local park
        points:
          type: integer
          example: 10

    PostcodeMetric:
      type: object
      required: [postcode, total_points]
      properties:
        postcode:
          type: string
          example: EH1 1YZ
        total_points:
          type: integer
          example: 250

    PostcodeDetailMetric:
      type: object
      required: [postcode, carbon_intensity, air_quality, users]
      properties:
        postcode:
          type: string
          example: EH1 1YZ
        carbon_intensity:
          type: number
          nullable: true
          description: Carbon intensity of electricity in gCO2/kWh (null if unavailable).
          example: 142.5
        air_quality:
          type: integer
          nullable: true
          description: Air quality index, 1 (low) to 10 (very high) (null if unavailable).
          minimum: 1
          maximum: 10
          example: 3
        users:
          type: array
          description: Users in the postcode, ordered by points descending.
          items:
            $ref: '#/components/schemas/UserMetric'

    UserMetric:
      type: object
      required: [user_id, name, points]
      properties:
        user_id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: Jane Smith
        points:
          type: integer
          example: 120

    Coordinates:
      type: object
      required: [postcode, latitude, longitude]
      properties:
        postcode:
          type: string
          example: EH1 1YZ
        latitude:
          type: number
          format: double
          minimum: -90
          maximum: 90
          example: 55.9533
        longitude:
          type: number
          format: double
          minimum: -180
          maximum: 180
          example: -3.1883

    Error:
      type: object
      required: [message]
      properties:
        message:
          type: string
          example: Resource not found

  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: Resource already exists
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'